Refactoring modules package and removing generated test files after test
